renamed test module to avoid naming conflict with pykafka
