+++ tests/pykafka_tests/rdkafka/test_sasl.py
