+++ tests/pykafka_tests/test_balancedconsumer.py
