+++ tests/pykafka_tests/test_cluster.py
