+++ tests/pykafka_tests/test_connection.py
