+++ tests/pykafka_tests/test_producer.py
