+++ tests/pykafka_tests/test_sasl.py
