+++ tests/pykafka_tests/test_simpleconsumer.py
