+++ tests/pykafka_tests/utils/test_compression.py
