+++ tests/pykafka_tests/utils/test_struct_helpers.py
